create api response model that contains new jwt response and moved api response common and check credential response
