+++ models/api_response_model/jwt_response.go
@@ -1,0 +1,9 @@
+package api_response_model
+
+import "time"
+
+type JwtResponse struct {
+	Code   int       `json:"code"`
+	Expire time.Time `json:"expire"`
+	Token  string    `json:"token"`
+}
